+++ src/main/java/com/chungbazi/server/domain/policy/dto/internal/ApplyPeriod.java
@@ -1,0 +1,14 @@
+package com.chungbazi.server.domain.policy.dto.internal;
+
+import com.chungbazi.server.domain.policy.enums.RecruitmentStatus;
+import com.chungbazi.server.domain.policy.enums.RecruitmentType;
+import java.time.LocalDate;
+
+public record ApplyPeriod(
+        LocalDate startDate,
+        LocalDate endDate,
+        String periodText,
+        RecruitmentType recruitmentType,
+        RecruitmentStatus recruitmentStatus
+) {
+}
