+++ src/main/java/com/chungbazi/server/domain/policy/dto/internal/DateRange.java
@@ -1,0 +1,9 @@
+package com.chungbazi.server.domain.policy.dto.internal;
+
+import java.time.LocalDate;
+
+public record DateRange(
+        LocalDate startDate,
+        LocalDate endDate
+) {
+}
