+++ src/main/java/com/chungbazi/server/domain/policy/dto/internal/ParsedPeriod.java
@@ -1,0 +1,7 @@
+package com.chungbazi.server.domain.policy.dto.internal;
+
+public record ParsedPeriod(
+        DateRange dateRange,
+        String periodText
+) {
+}
